# api_doc/core_components/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AuroraAI core components API",
    "description": "API endpoints implemented by AuroraAI core components",
    "version": "0.0.1",
    "contact": {
      "name": "AuroraAI core components",
      "url": "https://auroraai.fi",
      "email": "auroraai@suomi.fi"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "AuroraAI core components"
    }
  ],
  "paths": {
...
    "/oauth/token": {
      "post": {
        "summary": "OAuth2.0 /oauth/token endpoint",
        "operationId": "post-oauth-token",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "token_type": {
                      "type": "string",
                      "enum": [
                        "bearer"
                      ],
                      "description": "always \"bearer\""
                    },
                    "expires_in": {
                      "type": "integer",
                      "description": "lifetime of the access token in seconds"
                    },
                    "refresh_token": {
                      "type": "string",
                      "description": "mandatory when grant_type is \"authorization_code\""
                    },
                    "scope": {
                      "type": "string",
                      "description": "scopes granted for the access token"
                    },
                    "id_token": {
                      "type": "string",
                      "description": "JWT id token of the subject"
                    }
                  },
                  "required": [
                    "access_token",
                    "token_type",
                    "expires_in",
                    "scope",
                    "id_token"
                  ]
                }
              }
            }
          }
        },
        "description": "OAuth 2.0 token endpoint to request id, access and refresh tokens.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "grant_type": {
                    "type": "string",
                    "enum": [
                      "authorization_code",
                      "refresh_token"
                    ],
                    "description": "use \"authorization_code\" when getting the tokens for the first time, \"refresh_token\" afterwards"
                  },
                  "code": {
                    "type": "string",
                    "minLength": 1,
                    "description": "mandatory when grant_type is \"authorization_code\""
                  },
                  "redirect_uri": {
                    "type": "string",
                    "description": "mandatory when grant_type is \"authorization code\"",
                    "format": "uri"
                  },
                  "refresh_token": {
                    "type": "string",
                    "description": "mandatory when grant_type is \"refresh_token\"",
                    "minLength": 1
                  },
                  "scope": {
                    "type": "string",
                    "description": "optionally reduce scope when grant_type is \"refresh_token\"",
                    "example": "age store:municipality_code"
                  }
                },
                "required": [
                  "grant_type"
                ]
              }
            }
          }
        },
        "security": [
          {
            "basic_client_id_client_secret": []
          }
        ],
        "tags": [
          "OAuth2.0"
        ]
      }
    },
    "/oauth/revoke": {
      "post": {
        "summary": "OAuth2.0 /oauth/revoke endpoint",
        "operationId": "post-oauth-revoke",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "description": "Revokes user's tokens associated with the given token (refresh or access token).",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 1
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "basic_client_id_client_secret": []
          }
        ],
        "tags": [
          "OAuth2.0"
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "basic_client_id_client_secret": {
        "type": "http",
        "scheme": "basic",
        "description": "Authenticate with client_id and client_secret"
      },
      "open_id_connect": {
        "type": "openIdConnect",
        "openIdConnectUrl": "/oauth/.well-known/openid-configuration"
      }
    }
  },
...
}